Rearranged the RayCast to make it seem more intuitive to use, also renamed the file
